implemented CircularFwdListConstIterator
